payment info and login styles
