+++ app/login/template.hbs
@@ -1,45 +1,40 @@
 {{#login-box}}
-  <div class="col-xs-6 col-xs-offset-3">
-    <h3 class="heading">
-      Log in to Your Aptible Account
-    </h3>
-    <div class="sub-heading">
-      Don't have an Aptible account?
-      {{link-to "Create Your Aptible Account" 'signup'}}
-    </div>
-    <div class="alerts-partial-wrapper">
+  <div class="row">
+    <div class="col-xs-6 col-xs-offset-3">
       {{#if model.error}}
-        <div class="alert alert-warning fade in">
+        <div class="alert alert-danger fade in">
           <a class="close" data-dismiss="alert">×</a>
           {{model.error}}
         </div>
       {{/if}}
-    </div>
+      <div class="panel panel-subdued panel-focused">
+        <div class="panel-heading">
+          <div class="panel-heading-link">
+            {{link-to 'Create an account' 'signup'}}
+          </div>
+          <h3>Sign in to your Aptible account</h3>
+        </div>
+        <div class="panel-body">
+          <div class="form-group">
+            <label>Email</label>
+            {{input type="email"
+              value=model.email
+              autofocus=true
+              class="form-control"
+              placeholder="Email"}}
+          </div>
+          <div class="form-group">
+            <label>Password</label>
+            {{input type="password"
+              value=model.password
+              class="form-control"
+              placeholder="Password"}}
+          </div>
+          <button class="btn btn-primary btn-block btn-lg" {{action "login" model}}>Log in</button>
 
-    <form {{action "login" on="submit"}}>
-      <div class="input-group">
-        <span class="input-group-addon">
-          <i class="fa fa-envelope"></i>
-        </span>
-        {{input type="email"
-            value=model.email
-            autofocus=true
-            class="form-control"
-            placeholder="Email"}}
-      </div>
-      <div class="input-group">
-        <span class="input-group-addon">
-          <i class="fa fa-lock"></i>
-        </span>
-        {{input type="password"
-            value=model.password
-            class="form-control"
-            placeholder="Password"}}
-      </div>
-      <div class="form-group">
-        <button class="btn btn-primary" {{action "login" model}}>Log in</button>
         <a class="pull-right forgot-password-link" href="#">Forgot your password?</a>
+        </div>
       </div>
-    </form>
+    </div>
   </div>
 {{/login-box}}
