+++ app/styles/components/login-box.scss
@@ -13,4 +13,15 @@
       margin: 15px 0;
     }
   }
+
+  .expiration-month {
+    width: 35%;
+    float: left;
+    margin-right: 10%;
+  }
+
+  .expiration-year {
+    width: 55%;
+    float: left;
+  }
 }
